backend/requirements.txt
alembic==1.7.5
fastapi==0.85.1
Jinja2==3.1.2
uvicorn==0.16.0
python-dotenv==0.21.0
asyncpg==0.26.0
psycopg2-binary
pytz==2022.4
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
pydantic==1.10.4
inflect==5.6.0
